requirements.txt
fastapi
uvicorn
pydantic
openai
sse-starlette
pyyaml